code refactoring & make prediction system
