+++ Model/RNN_training.py
@@ -1,0 +1,124 @@
+import tensorflow as tf
+import numpy as np
+from tensorflow import keras
+from tensorflow.keras.preprocessing.sequence import pad_sequences
+
+def preprocess(sequences, max_len, dic, mode='source'):
+    assert mode in ['source', 'target'], 'source와 target 중에 선택'
+
+    if mode == 'source':
+        # preprocessing for source (use in encoder)
+        s_input = list(map(lambda sentence: [dic.get(token) for token in sentence], sequences))
+        s_len = list(map(lambda sentence: len(sentence), s_input))
+        # source의 길이는 3으로 고정될 것이므로 pad_sequences가 필요없어서 주석처리
+        s_input = pad_sequences(sequences=s_input, maxlen=max_len, padding='post', truncating='post')
+        return s_len, s_input
+
+    elif mode == 'target':
+        # preprocessing for target (use in decoder)
+        # decoder input
+        # bos: beginning of sentence eos: end of sentence
+        t_input = list(map(lambda sentence: ['<bos>'] + sentence + ['<eos>'], sequences))
+        t_input = list(map(lambda sentence: [dic.get(token) for token in sentence], t_input))
+        t_len = list(map(lambda sentence: len(sentence), t_input))
+        t_input = pad_sequences(sequences=t_input, maxlen=max_len, padding='post',
+                                truncating='post')  # truncating: 길이가 초과할경우 어디서부터 유효할지
+
+        # decoder output
+        t_output = list(map(lambda sentence: sentence + ['<eos>'], sequences))
+        t_output = list(map(lambda sentence: [dic.get(token) for token in sentence], t_output))
+        t_output = pad_sequences(sequences=t_output, maxlen=max_len, padding='post', truncating='post')
+
+        return t_len, t_input, t_output
+
+
+class Encoder(tf.keras.Model):
+    def __init__(self, vocab_size, embedding_dim, enc_units, batch_sz):
+        super(Encoder, self).__init__()
+        self.batch_sz = batch_sz
+        self.enc_units = enc_units
+        self.embedding = tf.keras.layers.Embedding(vocab_size, embedding_dim)
+        self.gru = tf.keras.layers.GRU(self.enc_units,
+                                       return_sequences=True,
+                                       return_state=True,
+                                       recurrent_activation='sigmoid',
+                                       recurrent_initializer='glorot_uniform'
+                                       )
+
+    def __call__(self, x, hidden):
+        x = self.embedding(x)
+        output, state = self.gru(x, initial_state=hidden)
+        return output, state
+
+    def initialize_hidden_state(self):
+        return tf.zeros((self.batch_sz, self.enc_units))
+
+
+class Decoder(tf.keras.Model):
+    def __init__(self, vocab_size, embedding_dim, dec_units, batch_sz):
+        super(Decoder, self).__init__()
+        self.batch_sz = batch_sz
+        self.dec_units = dec_units
+        self.embedding = tf.keras.layers.Embedding(vocab_size, embedding_dim)
+        self.gru = tf.keras.layers.GRU(self.dec_units,
+                                       return_sequences=True,
+                                       return_state=True,
+                                       recurrent_activation='sigmoid',
+                                       recurrent_initializer='glorot_uniform'
+                                       )
+        self.fc = tf.keras.layers.Dense(vocab_size)
+
+        # used for attention
+        self.W1 = tf.keras.layers.Dense(self.dec_units)
+        self.W2 = tf.keras.layers.Dense(self.dec_units)
+        self.V = tf.keras.layers.Dense(1)
+
+    def __call__(self, x, hidden, enc_output):  # 너무 어려운 개념이 많이쓰임... attention 눙물나쥬
+        # enc_output shape == (batch_size, max_length, hidden_size)
+
+        # hidden shape == (batch_size, hidden size)
+        # hidden_with_time_axis shape == (batch_size, 1, hidden size)
+        # we are doing this to perform addition to calculate the score
+        hidden_with_time_axis = tf.expand_dims(hidden, 1)
+        # * `score = FC(tanh(FC(EO) + FC(H)))`
+        # score shape == (batch_size, max_length, 1)
+        # we get 1 at the last axis because we are applying tanh(FC(EO) + FC(H)) to self.V
+        score = self.V(tf.nn.tanh(self.W1(enc_output) + self.W2(hidden_with_time_axis)))
+
+        # * `attention weights = softmax(score, axis = 1)`. Softmax by default is applied on the last axis but here we want to apply it on the *1st axis*, since the shape of score is *(batch_size, max_length, 1)*. `Max_length` is the length of our input. Since we are trying to assign a weight to each input, softmax should be applied on that axis.
+        # attention_weights shape == (batch_size, max_length, 1)
+        attention_weights = tf.nn.softmax(score, axis=1)
+
+        # context_vector shape after sum == (batch_size, hidden_size)
+        # * `context vector = sum(attention weights * EO, axis = 1)`. Same reason as above for choosing axis as 1.
+        context_vector = attention_weights * enc_output
+        context_vector = tf.reduce_sum(context_vector, axis=1)
+
+        # x shape after passing through embedding == (batch_size, 1, embedding_dim)
+        # * `embedding output` = The input to the decoder X is passed through an embedding layer.
+        x = self.embedding(x)
+
+        # x shape after concatenation == (batch_size, 1, embedding_dim + hidden_size)
+        # * `merged vector = concat(embedding output, context vector)`
+        x = tf.concat([tf.expand_dims(context_vector, 1), x], axis=-1)
+
+        # passing the concatenated vector to the GRU
+        output, state = self.gru(x)
+
+        # output shape == (batch_size * 1, hidden_size)
+        output = tf.reshape(output, (-1, output.shape[2]))
+
+        # output shape == (batch_size * 1, vocab)
+        x = self.fc(output)
+
+        return x, state, attention_weights
+
+    def initialize_hidden_state(self):
+        return tf.zeros((self.batch_sz, self.dec_units))
+
+def loss_function(real, pred):
+    mask = 1 - np.equal(real, 0)
+    loss_ = tf.nn.sparse_softmax_cross_entropy_with_logits(labels=real, logits=pred) * mask
+
+    return tf.reduce_mean(loss_)
+
