+++ Model/data_out/test_input.txt
@@ -24,7 +24,7 @@
 삼성 LG 현대
 대기업 총수 청문회
 조국 차명계좌 논란
-청문회 이번에도 졸속
+청문회 이번에도 파행
 한미 군사훈련 연기
 文정부 지지율 하락
 최순실 징역 선고
@@ -90,13 +90,13 @@ GM 군산 공장
 아파트 가격 껑충
 셀트리온 주식 폭락
 일본 대기업 불매
-52시간제 적용 부작용
-소비세 신설 반대
+52시간 적용 부작용
+소비자 지수 반대
 GM 실적 최대
 블랙리스트 등록 박원순
 법무부 장관 선임
 이명박 문건 발견
-北 핵시설 건설
+북한 핵시설 건설
 우병우 해외 도피
