--- Naver_News_Title_Crawler/run.py
@@ -1,14 +1,0 @@
-from Naver_News_Title_Crawler import downloader
-
-
-xls_path = 'data.xlsx'
-date = '20190816'
-num = 25000
-
-
-downloader.download(num, xls_path, date)
-
-
-
-
-
